Bug fix.
Fixed a bug that caused the display of an invisible dialog box (a dark screen when starting the application).

Changed the text color of the registration button and the button for changing the visibility of the password.

Connected service http://ip-api.com/ to verify the existence of a domain name.